Homework13 (ADO.NET) Add CRUD operations on products table in disconnected mode, add fetch all products, add bulk delete of orders

+++ 13-ADO.NET/DB_Operations/Models/ProductModel.cs
@@ -8,9 +8,9 @@ namespace DB_Operations.Models
     {
         public string Name { get; set; }
         public string Description { get; set; }
-        public float Weight { get; set; }
-        public float Height { get; set; }
-        public float Width { get; set; }
-        public float Length { get; set; }
+        public decimal Weight { get; set; }
+        public decimal Height { get; set; }
+        public decimal Width { get; set; }
+        public decimal Length { get; set; }
     }
 }
